+++ src/Telegram.Bot/Enums/Enums.cs
@@ -1,0 +1,11 @@
+﻿//// ReSharper disable once CheckNamespace
+//
+//namespace Telegram.Bot
+//{
+//    /// <summary>
+//    /// Provides constant values as defined in the Bot API
+//    /// </summary>
+//    public static partial class Enums
+//    {
+//    }
+//}