+++ test/Telegram.Bot.Tests.Integ/Framework/TestsFixture.cs
@@ -28,12 +28,15 @@ namespace Telegram.Bot.Tests.Integ.Framework
 
         public RunSummary RunSummary { get; } = new RunSummary();
 
+        public static TestsFixture Instance { get; private set; }
+
         private CancellationToken CancellationToken =>
             new CancellationTokenSource(TimeSpan.FromSeconds(45)).Token;
 
         public TestsFixture()
         {
             InitAsync().GetAwaiter().GetResult();
+            Instance = this;
         }
 
         public void Dispose()
@@ -56,10 +59,33 @@ namespace Telegram.Bot.Tests.Integ.Framework
             ).GetAwaiter().GetResult();
         }
 
+        public Task<Message> SendTestInstructionsAsync(
+            string instructions,
+            ChatId chatid = default,
+            bool startInlineQuery = default
+        )
+        {
+            string text = string.Format(Constants.InstructionsMessageFormat, instructions);
+            chatid = chatid ?? SupergroupChat.Id;
+
+            IReplyMarkup replyMarkup = startInlineQuery
+                ? (InlineKeyboardMarkup) InlineKeyboardButton.WithSwitchInlineQueryCurrentChat("Start inline query")
+                : default;
+
+            return BotClient.SendTextMessageAsync(
+                chatid,
+                text,
+                ParseMode.Markdown,
+                replyMarkup: replyMarkup,
+                cancellationToken: CancellationToken
+            );
+        }
+
         public async Task<Message> SendTestCaseNotificationAsync(string testcase,
             string instructions = default,
             ChatId chatid = default,
-            bool startInlineQuery = default)
+            bool startInlineQuery = default
+        )
         {
             Message msg = await SendNotificationToChatAsync(false, testcase, instructions, chatid, startInlineQuery);
             return msg;
@@ -139,8 +165,13 @@ namespace Telegram.Bot.Tests.Integ.Framework
 #endif
         }
 
-        private Task<Message> SendNotificationToChatAsync(bool isForCollection, string name,
-            string instructions = default, ChatId chatid = default, bool switchInlineQuery = default)
+        private Task<Message> SendNotificationToChatAsync(
+            bool isForCollection,
+            string name,
+            string instructions = default,
+            ChatId chatid = default,
+            bool switchInlineQuery = default
+        )
         {
             var textFormat = isForCollection
                 ? Constants.StartCollectionMessageFormat