+++ test/Telegram.Bot.Tests.Integ/Sending Messages/TextMessageTests.cs
@@ -27,20 +27,16 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
             _classFixture = classFixture;
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldSendTextMessage)]
+        [OrderedFact("Should send text message")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
         public async Task Should_Send_Text_Message()
         {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldSendTextMessage);
-
-            const string text = "Hello world!";
-
             Message message = await BotClient.SendTextMessageAsync(
                 chatId: _fixture.SupergroupChat.Id,
-                text: text
+                text: "Hello world!"
             );
 
-            Assert.Equal(text, message.Text);
+            Assert.Equal("Hello world!", message.Text);
             Assert.Equal(MessageType.Text, message.Type);
             Assert.Equal(_fixture.SupergroupChat.Id.ToString(), message.Chat.Id.ToString());
             Assert.InRange(message.Date, DateTime.UtcNow.AddSeconds(-10), DateTime.UtcNow.AddSeconds(2));
@@ -49,12 +45,10 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
             ));
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldSendTextMessageToChannel)]
+        [OrderedFact("Should send text message to channel")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
         public async Task Should_Send_Text_Message_To_Channel()
         {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldSendTextMessageToChannel);
-
             string text = $"Hello members of channel {_classFixture.ChannelChatId}";
 
             Message message = await BotClient.SendTextMessageAsync(
@@ -68,11 +62,14 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
             Assert.Equal(_classFixture.ChannelChat.Username, message.Chat.Username);
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldForwardMessage)]
+        [OrderedFact("Should forward a message to same chat")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.ForwardMessage)]
         public async Task Should_Forward_Message()
         {
-            Message message1 = await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldForwardMessage);
+            Message message1 = await BotClient.SendTextMessageAsync(
+                /* chatId: */ _fixture.SupergroupChat,
+                /* text: */ "➡️ Message to be forwared ⬅️"
+            );
 
             Message message2 = await BotClient.ForwardMessageAsync(
                 chatId: _fixture.SupergroupChat,
@@ -92,12 +89,11 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
             );
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldParseMarkDownEntities)]
+        [OrderedFact("Should send markdown formatted text message and parse its entities. " +
+                     "Link preview should not appear.")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
         public async Task Should_Parse_MarkDown_Entities()
         {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldParseMarkDownEntities);
-
             const string url = "https://telegram.org/";
             Dictionary<MessageEntityType, string> entityValueMappings = new Dictionary<MessageEntityType, string>
             {
@@ -127,14 +123,13 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
             ));
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldParseHtmlEntities)]
+        [OrderedFact("Should send HTML formatted text message and parse its entities. Link preview should not appear.")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
         public async Task Should_Parse_HTML_Entities()
         {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldParseHtmlEntities);
-
             const string url = "https://telegram.org/";
-            (MessageEntityType Type, string Value)[] entityValueMappings = {
+            (MessageEntityType Type, string Value)[] entityValueMappings =
+            {
                 (MessageEntityType.Bold, "<b>bold</b>"),
                 (MessageEntityType.Bold, "<strong>&lt;strong&gt;</strong>"),
                 (MessageEntityType.Italic, "<i>italic</i>"),
@@ -166,13 +161,12 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
             ));
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldParseMessageEntitiesIntoValues)]
+        [OrderedFact("Should send text message and parse its entity values")]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
         public async Task Should_Parse_Message_Entities_Into_Values()
         {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldParseMessageEntitiesIntoValues);
-
-            (MessageEntityType Type, string Value)[] entityValueMappings = {
+            (MessageEntityType Type, string Value)[] entityValueMappings =
+            {
                 (MessageEntityType.PhoneNumber, "+386 12 345 678"),
                 (MessageEntityType.Cashtag, "$EUR"),
                 (MessageEntityType.Hashtag, "#TelegramBots"),
@@ -193,24 +187,6 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
                 message.Entities.Select(e => e.Type)
             );
             Assert.Equal(entityValueMappings.Select(t => t.Value), message.EntityValues);
-        }
-
-        private static class FactTitles
-        {
-            public const string ShouldSendTextMessage = "Should send text message";
-
-            public const string ShouldSendTextMessageToChannel = "Should send text message to channel";
-
-            public const string ShouldForwardMessage = "Should forward a message to same chat";
-
-            public const string ShouldParseMarkDownEntities =
-                "Should send markdown formatted text message and parse its entities. Link preview should not appear.";
-
-            public const string ShouldParseHtmlEntities =
-                "Should send HTML formatted text message and parse its entities. Link preview should not appear.";
-
-            public const string ShouldParseMessageEntitiesIntoValues =
-                "Should send text message and parse its entity values";
         }
 
         public class Fixture : ChannelChatFixture